chore: move archer mocks to archer dir
