+++ internal/pkg/cli/pipeline_init_test.go
@@ -9,9 +9,9 @@ import (
 	"os"
 	"testing"
 
+	archermocks "github.com/aws/amazon-ecs-cli-v2/internal/pkg/archer/mocks"
 	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/aws/secretsmanager"
 	climocks "github.com/aws/amazon-ecs-cli-v2/internal/pkg/cli/mocks"
-	archermocks "github.com/aws/amazon-ecs-cli-v2/mocks"
 	"github.com/gobuffalo/packd"
 	"github.com/golang/mock/gomock"
 	"github.com/spf13/afero"
